Add general tabs at the bottom of the project list.
These are sunk to the bottom of the column, but if the column is too
tall then the whole column scrolls without anything overlapping.
This behaviour is largely lifted from http://www.cssstickyfooter.com/
though some of the dimensions have some tweaks so that the borders
at the join point collapse to a single width, and of course it's all
been extracted into being based on variables thanks to SASS.

If we end up dynamically modifying the list at the bottom of the column
then something else will need to be used, since unfortunately the
height of those tabs is hard-coded into *other* parts of the page.

Tested in Fx 7 and Chrome 15 as working, mostly works in Opera 11.
The original site has a fix for Opera, but I couldn't get it to work
in this adaptation.