Add various other "modes" that I think we're going to want.
I'm not sure how we're going to launch a mode yet. I think they
could be from tabs on the LHS, but this might encroach on space for
projects if we had more than a few (which we probably will).

Each of these is very much an outline with comments and ideas in
place too. Some have implementations or thought towards that.